app-webhook-orchestrator/requirements.txt
aiohttp            == 3.8.4
python-socketio    == 5.8.0
PyYAML             == 6.0